requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
dotenv==0.0.5
dpath==2.0.6
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.6
packaging==21.3
paho-mqtt==1.6.1
pluggy==1.0.0
py==1.11.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.1
PyYAML==6.0
requests==2.26.0
RPi.GPIO==0.7.0
scipy==1.7.3
simpful==2.6.2
tomli==2.0.1
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
yapf==0.32.0
zipp==3.6.0
